migration: Extracted precheck functionality
